requirements.txt
pdfplumber==0.11.4
langchain-text-splitters==0.3.6
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
google-generativeai==0.8.4
openai==1.65.5
streamlit==1.42.2
python-dotenv==1.0.1
numpy>=2.0.0
